+++ plugins/corerouter/cr_map.c
@@ -19,7 +19,7 @@ int uwsgi_cr_map_use_cache(struct uwsgi_corerouter *ucr, struct corerouter_peer 
 }
 
 int uwsgi_cr_map_use_pattern(struct uwsgi_corerouter *ucr, struct corerouter_peer *peer) {
-	int tmp_socket_name_len = 0;
+	size_t tmp_socket_name_len = 0;
 	ucr->magic_table['s'] = uwsgi_concat2n(peer->key, peer->key_len, "", 0);
 	peer->tmp_socket_name = magic_sub(ucr->pattern, ucr->pattern_len, &tmp_socket_name_len, ucr->magic_table);
 	free(ucr->magic_table['s']);
